+++ src/main/java/com/github/danielwegener/logback/kafka/keying/ContextNameKeyingStrategy.java
@@ -11,8 +11,9 @@ import java.nio.ByteBuffer;
  * This strategy uses logbacks CONTEXT_NAME to partition the log messages to kafka.
  * This is ensures that all log messages logged by the same logging context will remain in the correct order for any consumer.
  * But this strategy can lead to uneven log distribution for a small number of hosts (compared to the number of partitions).
+ * @since 0.0.1
  */
-public class ContextNameKeyingStrategy extends ContextAwareBase implements KeyingStrategy {
+public class ContextNameKeyingStrategy extends ContextAwareBase implements KeyingStrategy<ILoggingEvent> {
 
     private byte[] contextNameHash = null;